Changed the primitive data type of properties 'price' and 'wallet' to double instead of int in Customer.java and Artwork.java. Made a @BeforeEach annotation in GalleryTest.java and two tests to verify the name and amount in the gallery's till. Used TDD to create the appropriate methods and getters in Gallery.java to pass the tests.

+++ Art_Gallery/src/main/java/Artwork.java
@@ -1,9 +1,9 @@
 public class Artwork {
     private String title;
     private Artist artist;
-    private int price;
+    private double price;
 
-    public Artwork(String title, Artist artist, int price) {
+    public Artwork(String title, Artist artist, double price) {
         this.title = title;
         this.artist = artist;
         this.price = price;
@@ -25,7 +25,7 @@ public class Artwork {
         this.artist = artist;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
